+++ src/constant/index.ts
@@ -294,4 +294,5 @@ export enum OVERVIEW_CARD {
   ProviderTrafficOverview = 'ProviderTrafficOverview',
   TopologyCharts = 'TopologyCharts',
   ConnectionHistory = 'ConnectionHistory',
+  RuleHitCountCard = 'RuleHitCountCard',
 }